+++ R/data.R
@@ -10,9 +10,9 @@
 
 #' @title Leonardo da Vinci's 72-sided sphere
 #'
-#' @description A matrix giving the 20 vertices of da Vinci's 72-sided sphere, 
+#' @description A matrix giving the 62 vertices of da Vinci's 72-sided sphere, 
 #'   a convex polyhedra with 72 faces.
 #'
-#' @format A matrix with 20 rows and 3 columns.
-#' @source \url{https://www.georgehart.com/virtual-polyhedra/leonardo.html}
+#' @format A matrix with 62 rows and 3 columns.
+#' @source \url{http://www.matematicasvisuales.com/english/html/geometry/space/sphereCampanus.html}
 "daVinciSphere"